Stage 5 WIP: types/expressions complete (decimal, temporal, CAST/CASE/LIKE/functions)
M2: exact DECIMAL arithmetic; DATE/TIME/DATETIME2/UNIQUEIDENTIFIER/VARBINARY;
CAST/CONVERT, CASE (simple+searched), LIKE (% _ [] [^] ESCAPE), IN, BETWEEN,
ISNULL/COALESCE/IIF/NULLIF, and a string/number/date function library. Numeric
implicit-conversion ladder; eval recursion-depth guard (191). New modules:
decimal, temporal, like, guid, functions.

+++ crates/truthdb-sql/src/temporal.rs
@@ -14,8 +14,8 @@ const DAYS_0001_TO_1970: i64 = 719_162;
 fn days_from_civil(y: i64, m: u32, d: u32) -> i64 {
     let y = if m <= 2 { y - 1 } else { y };
     let era = if y >= 0 { y } else { y - 399 } / 400;
-    let yoe = (y - era * 400) as i64;
-    let doy = ((153 * (if m > 2 { m - 3 } else { m + 9 }) as i64 + 2) / 5 + d as i64 - 1) as i64;
+    let yoe = y - era * 400;
+    let doy = (153 * (if m > 2 { m - 3 } else { m + 9 }) as i64 + 2) / 5 + d as i64 - 1;
     let doe = yoe * 365 + yoe / 4 - yoe / 100 + doy;
     era * 146097 + doe - 719468
 }
